requirements.txt
Django
djangorestframework
django-simple-history
